lms_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
scikit-learn
joblib
gunicorn
psycopg2-binary
